+++ CHANGELOG.md
@@ -2,9 +2,13 @@
 
 ## [Unreleased]
 
-* Add `aws-sso-cli completion --source` flag to generate completion script and
-  print to stdout. #779
-* UrlExecCommand now supports commands in `~` and the `$HOME` environment variable. #816
+## [v1.15.0] - 2024-04-05
+
+### New Features
+
+ * Add `aws-sso-cli completion --source` flag to generate completion script and
+   print to stdout. #779
+ * UrlExecCommand now supports commands in `~` and the `$HOME` environment variable. #816
 
 ## [v1.14.3] - 2024-01-15